+++ src/Pages/Accounts/Components/AccountButtons/Components/AccountAssets.tsx
@@ -1,20 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import styled from 'styled-components';
 import { useAssets, useAccounts } from 'redux/hooks';
-import { Accordion, Table } from 'Components';
+import { Table } from 'Components';
 
-const ButtonWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  height: 100%;
-  width: 100%;
-  overflow: hidden;
-`;
-
-const AccountAssets = () => {
+export const AccountAssets = () => {
   const [tableCurrentPage, setTableCurrentPage] = useState(1);
-  const { addressId } = useParams();
+  const { addressId } = useParams<{ addressId: string }>();
   const {
     accountAssets,
     accountAssetsTotal,
@@ -50,25 +41,14 @@ const AccountAssets = () => {
   ];
 
   return (
-    <ButtonWrapper>
-      <Accordion
-        showChevron
-        title={`Assets (${accountAssetsTotal})`}
-        titleFont={`font-weight: bold; font-size: 1.4rem`}
-        startOpen={true}
-        dontDrop={true}
-      >
-        <Table
-          changePage={setTableCurrentPage}
-          currentPage={tableCurrentPage}
-          isLoading={tableLoading}
-          tableData={tableData}
-          tableHeaders={tableHeaders}
-          totalPages={tablePages}
-        />
-      </Accordion>
-    </ButtonWrapper>
+    <Table
+      changePage={setTableCurrentPage}
+      currentPage={tableCurrentPage}
+      isLoading={tableLoading}
+      tableData={tableData}
+      tableHeaders={tableHeaders}
+      totalPages={tablePages}
+      title={`Total Assets: ${accountAssetsTotal}`}
+    />
   );
 };
-
-export default AccountAssets;