+++ src/Pages/Accounts/Components/AccountButtons/Components/AccountDelegations.tsx
@@ -1,0 +1,55 @@
+import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
+import { useValidators, useAccounts } from 'redux/hooks';
+import { Table } from 'Components';
+import { formatDenom } from 'utils';
+
+export const AccountDelegations = () => {
+  const [tableCurrentPage, setTableCurrentPage] = useState(1);
+  const [tableData, setTableData] = useState<any[]>([]);
+  const {
+    accountDelegations,
+    accountDelegationsTotal: { amount, denom },
+    accountDelegationsLoading,
+    accountDelegationsPages: tablePages,
+    getAccountDelegations,
+  } = useAccounts();
+  const { allValidators, allValidatorsLoading, getAllValidators } = useValidators();
+  const { addressId } = useParams<{ addressId: string }>();
+  const tableCount = 5;
+
+  useEffect(() => {
+    // pulling first 100 validators with status=all
+    getAllValidators({});
+    getAccountDelegations({ address: addressId, page: tableCurrentPage, count: tableCount });
+  }, [addressId, getAllValidators, getAccountDelegations, tableCurrentPage]);
+
+  useEffect(() => {
+    setTableData(
+      accountDelegations.map((d) => {
+        const validator = allValidators.find((v) => v.addressId === d.validatorSrcAddr);
+        return { ...validator, ...d };
+      })
+    );
+  }, [allValidators, accountDelegations, setTableData]);
+
+  const totalDelegations = formatDenom(Number(amount), denom, { decimal: 2 });
+
+  const tableHeaders = [
+    //{ displayName: 'Staking', dataName: 'manageStaking' },
+    { displayName: 'Moniker', dataName: 'moniker' },
+    { displayName: 'Amount', dataName: 'amount' },
+  ];
+
+  return (
+    <Table
+      changePage={setTableCurrentPage}
+      currentPage={tableCurrentPage}
+      isLoading={accountDelegationsLoading || allValidatorsLoading}
+      tableData={tableData}
+      tableHeaders={tableHeaders}
+      totalPages={tablePages}
+      title={`Total Delegations: ${totalDelegations}`}
+    />
+  );
+};