+++ src/Pages/Accounts/Components/AccountButtons/Components/AccountDelegationsOwner.js
@@ -1,27 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import styled from 'styled-components';
-import { formatDenom } from 'utils';
 import { useValidators, useApp, useAccounts, useStaking } from 'redux/hooks';
-import { Accordion, Table } from 'Components';
+import { Table } from 'Components';
+import { formatDenom } from 'utils';
 import ManageStakingModal from '../../../../Validators/Components/ManageStakingModal';
 
-const ButtonWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  height: 100%;
-  width: 100%;
-  overflow: hidden;
-`;
-
-const AccountDelegationsOwner = () => {
+export const AccountDelegationsOwner = () => {
   const [tableCurrentPage, setTableCurrentPage] = useState(1);
   const [tableData, setTableData] = useState([]);
   const {
     accountDelegations,
+    accountDelegationsTotal: { amount, denom },
     accountDelegationsLoading,
     accountDelegationsPages: tablePages,
-    accountDelegationsTotal: { amount, denom },
     getAccountDelegations,
   } = useAccounts();
   const { addressId } = useParams();
@@ -32,18 +23,20 @@ const AccountDelegationsOwner = () => {
 
   useEffect(() => {
     // pulling first 100 validators with status=all
-    if (isLoggedIn) getAllValidators();
+    if (isLoggedIn) getAllValidators({});
     getAccountDelegations({ address: addressId, page: tableCurrentPage, count: tableCount });
   }, [isLoggedIn, addressId, getAllValidators, getAccountDelegations, tableCurrentPage]);
 
   useEffect(() => {
     setTableData(
-      accountDelegations.map(d => {
-        const validator = allValidators.find(v => v.addressId === d.validatorSrcAddr);
+      accountDelegations.map((d) => {
+        const validator = allValidators.find((v) => v.addressId === d.validatorSrcAddr);
         return { ...validator, ...d };
       })
     );
   }, [allValidators, accountDelegations, setTableData]);
+
+  const totalDelegations = formatDenom(Number(amount), denom, { decimal: 2 });
 
   const tableHeaders = [
     { displayName: 'Staking', dataName: 'manageStaking' },
@@ -51,36 +44,26 @@ const AccountDelegationsOwner = () => {
     { displayName: 'Amount', dataName: 'amount' },
   ];
 
-  const totalAmount = formatDenom(amount, denom, { decimal: 2 });
-
   return (
-    <ButtonWrapper>
-      <Accordion
-        showChevron
-        title={`Delegations (${totalAmount})`}
-        titleFont={`font-weight: bold; font-size: 1.4rem`}
-        dontDrop
-      >
-        <Table
-          changePage={setTableCurrentPage}
-          currentPage={tableCurrentPage}
-          isLoading={accountDelegationsLoading || allValidatorsLoading}
-          ManageStakingBtn={ManageStakingBtn}
-          tableData={tableData}
-          tableHeaders={tableHeaders}
-          totalPages={tablePages}
-        />
-        <ManageStakingModal
-          isDelegate={isDelegate}
-          isLoggedIn={isLoggedIn}
-          modalOpen={modalFns.modalOpen}
-          onClose={modalFns.deactivateModalOpen}
-          onStaking={handleStaking}
-          validator={validator || {}}
-        />
-      </Accordion>
-    </ButtonWrapper>
+    <>
+      <Table
+        changePage={setTableCurrentPage}
+        currentPage={tableCurrentPage}
+        isLoading={accountDelegationsLoading || allValidatorsLoading}
+        ManageStakingBtn={ManageStakingBtn}
+        tableData={tableData}
+        tableHeaders={tableHeaders}
+        totalPages={tablePages}
+        title={`Total Delegations: ${totalDelegations}`}
+      />
+      <ManageStakingModal
+        isDelegate={isDelegate}
+        isLoggedIn={isLoggedIn}
+        modalOpen={modalFns.modalOpen}
+        onClose={modalFns.deactivateModalOpen}
+        onStaking={handleStaking}
+        validator={validator || {}}
+      />
+    </>
   );
 };
-
-export default AccountDelegationsOwner;