+++ src/Pages/Accounts/Components/AccountButtons/Components/index.js
@@ -1,6 +1,0 @@
-export { default as AccountAssets } from './AccountAssets';
-export { default as AccountDelegations } from './AccountDelegations';
-export { default as AccountDelegationsOwner } from './AccountDelegationsOwner';
-export { default as AccountRewards } from './AccountRewards';
-export { default as AccountUnbondings } from './AccountUnbondings';
-export { default as AccountAttributes } from './AccountAttributes';