+++ src/Pages/Accounts/Components/index.ts
@@ -1,3 +1,3 @@
 export { default as AccountSpotlight } from './AccountSpotlight';
 export { default as AccountTxs } from './AccountTxs';
-export { default as AccountButtons } from './AccountButtons';
+export * from './AccountButtons';